pyproject.toml
[tool.poetry]
name = "fastapipractice"
version = "0.1.0"
description = ""
authors = ["Jang JiWoo <77304817+cheeou@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
bcrypt = "^4.2.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
webtool = "^0.0.11"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
